frontend: extract address resolver functions
